Updated for gaibu refactor
