+++ pom.xml
@@ -17,7 +17,7 @@
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <java.version>1.8</java.version>
         <suripu.version>0.8.5656</suripu.version>
-        <gaibu.version>0.1.31</gaibu.version>
+        <gaibu.version>0.1.37-SNAPSHOT</gaibu.version>
         <os-maven-plugin.version>1.4.1.Final</os-maven-plugin.version>
         <build-helper-maven-plugin.version>1.9.1</build-helper-maven-plugin.version>
     </properties>
